Extract RunStats into an independent class
The .hpp format was chosen because the class is tiny and there would be no benefit in splitting into .h and .cpp
